Create application component hierarchy and styling (#20)
## Description
This pull request removes the `.babelrc` file from the project as it is no longer needed.

## What changed
The `.babelrc` file was deleted. This file was previously used to specify how Babel, a JavaScript compiler, should behave. However, it is no longer necessary in our project.

## How to test
Since this is a deletion of a file, there are no direct features to test. However, you should ensure that the project still builds and runs correctly without this change. If there are any issues, they may be related to this change.

## Why make this change
The `.babelrc` file was initially added to configure Babel for our project. However, with recent updates and changes in our project setup, this file is no longer necessary. Removing it helps to keep our project clean and free of unnecessary configuration files.